+++ CHANGELOG.md
@@ -4,6 +4,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to
 [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- Fixed a bug where assembly output used a Unicode minus sign for negative
+  numbers on some locales, causing assembly errors.
+
 ## [1.6] (March 1, 2026)
 
 ### Added